proposal: Move keySetStringF to the hosts plugin
