+++ src/plugins/hosts/keymetaformatting.h
@@ -16,5 +16,6 @@
 void elektraAddLineComment (KeySet * comments, size_t spaces, const char * commentStart, const char * comment);
 void elektraAddInlineComment (KeySet * comments, size_t spaces, const char * commentStart, const char * comment);
 size_t elektraCountStartSpaces (const char * line);
+ssize_t keySetStringF (Key * key, const char * format, ...);
 
 #endif /* KEYMETAFORMATTING_H_ */
